renderer update

+++ src/miscellaneous/vec2.cpp
@@ -1,6 +1,6 @@
 #include "vec2.h"
 
-Vec2::Vec2(double _x, double _y): x(_x), y(_y) {}
+Vec2::Vec2(float _x, float _y): x(_x), y(_y) {}
 
 Vector2 Vec2::toRaylib() {
     return {x, y};
@@ -18,7 +18,7 @@ Vec2& Vec2::operator-=(const Vec2& v) {
     return *this;
 }
 
-Vec2& Vec2::operator*=(const double& k) {
+Vec2& Vec2::operator*=(const float& k) {
     x *= k;
     y *= k;
     return *this;
@@ -32,10 +32,10 @@ Vec2 operator-(const Vec2& v1, const Vec2& v2) {
     return Vec2(v1.x - v2.x, v1.y - v2.y);
 }
 
-Vec2 operator*(const Vec2& v, const double& k) {
+Vec2 operator*(const Vec2& v, const float& k) {
     return Vec2(v.x * k, v.y * k);
 }
 
-Vec2 operator*(const double& k, const Vec2& v) {
+Vec2 operator*(const float& k, const Vec2& v) {
     return Vec2(v.x * k, v.y * k);
 }
